Improve record reading performance
Read a complete record worth of data and construct a DataInputStream out of ByteArrayInputStream. Reads are then executed against the DataInputStream based on the underlying byte array instead of the RandomAccessFile which causes system calls to be executed and considerably slows down reading records. Further improvements could be made through buffering an entire chunck of file while reading through the Iterator.